refactor(nucleus-node): extract Firecracker config to firecracker_config.rs
Move 292 lines of Firecracker VM configuration structs, boot arg
injection, seccomp verification, and drive config helpers from
main.rs (3592→3304 lines) to a dedicated firecracker_config.rs module.

Add line ratchet for nucleus-node/src/main.rs:
ceiling 3304, target 2000, step 1 per merge.

No behavioral changes — pure extraction with module-qualified call sites.
